Handle toggling with +org/dwim-at-point
This fixes a bug wherein toggling images with enter or +org/dwim would
result in left over lines.

+org/dwim-at-point calls +org--toggle-inline-images-in-subtree

like org-toggle-inline-images, this function needs to be overridden so
that the sliced images are removed properly

This commit modifies +org--toggle-inline-images-in-subtree to call the
org-sliced-images-remove-inline-images function instead
of (delete-overlay ov). As of testing this, resolves the bug.